renamed dependency conflict package
