+++ test/Scripts.Integration.Test/DependencyConflictPackage/DependencyConflict/Runtime/System.Collections.Immutable.dll.meta
