+++ test/Scripts.Integration.Test/DependencyConflictPackage/DependencyConflict/Runtime/System.Numerics.Vectors.dll.meta
