+++ test/Scripts.Integration.Test/DependencyConflictPackage/DependencyConflict/Runtime/System.Text.Json.dll.meta
