+++ test/Scripts.Integration.Test/DependencyConflictPackage/DependencyConflict/Runtime/io.sentry.dependency-conflict.runtime.asmdef
@@ -7,7 +7,7 @@
     "allowUnsafeCode": false,
     "overrideReferences": false,
     "precompiledReferences": [
-        "DependencyConflict.dll",
+        "DependencyConflictPackage.dll",
         "System.Text.Json.dll",
         "System.Collections.Immutable.dll",
         "Microsoft.Bcl.AsyncInterfaces.dll"
