+++ test/Scripts.Integration.Test/DependencyConflictPackage/DependencyConflict/package.json.meta
