+++ test/Scripts.Integration.Test/DependencyConflictPackage/Directory.Build.targets
