+++ test/Scripts.Integration.Test/globals.ps1
@@ -85,9 +85,9 @@ $PackageReleaseOutput = "$(ProjectRoot)/test-package-release"
 $PackageReleaseAssetsPath = "$PackageReleaseOutput/Samples~/unity-of-bugs"
 
 # DependencyConflict: the assembly-aliasing stress-test fixture. The csproj
-# builds its plain (unaliased) System.*/Microsoft.* DLLs into package/Runtime,
-# and the package/ folder is copied into the test project as an embedded package.
-$DependencyConflictPackagePath = "$IntegrationScriptsPath/DependencyConflict/package"
+# builds its plain (unaliased) System.*/Microsoft.* DLLs into DependencyConflict/Runtime,
+# and the DependencyConflict/ folder is copied into the test project as an embedded package.
+$DependencyConflictPackagePath = "$IntegrationScriptsPath/DependencyConflictPackage/DependencyConflict"
 
 function FormatUnityPath
 {
